moves guice into it's own module and updates/migrates the README
guice will now be inside it's own module
README will now be converted to markdown to have the same
doc language in all projects
