+++ play-mailer-guice/src/main/resources/reference.conf
@@ -1,0 +1,4 @@
+play.modules {
+  enabled += "play.api.libs.mailer.MailerModule"
+  enabled += "play.api.libs.mailer.SMTPConfigurationModule"
+}
