+++ play-mailer-guice/src/main/scala/play/api/libs/mailer/MailerConfigurationModule.scala
