+++ play-mailer-guice/src/main/scala/play/api/libs/mailer/MailerModule.scala
