+++ play-mailer-guice/src/main/scala/play/api/libs/mailer/SMTPConfigurationModule.scala
