+++ play-mailer-guice/src/test/scala/play/api/libs/mailer/ConfigModule.scala
@@ -8,7 +8,9 @@ import play.api.{ Configuration, Environment }
  * Config Module to provide a shim for Play 2.5.x
  */
 class ConfigModule extends Module {
+
   override def bindings(environment: Environment, configuration: Configuration): Seq[Binding[_]] = Seq(
     bind[Config].toInstance(configuration.underlying)
   )
+
 }
