+++ play-mailer/src/main/java/play/libs/mailer/Email.java
