+++ play-mailer/src/main/resources/reference.conf
@@ -1,8 +1,4 @@
 play {
-  modules {
-    enabled += "play.api.libs.mailer.MailerModule"
-    enabled += "play.api.libs.mailer.SMTPConfigurationModule"
-  }
   mailer {
     host = null
     port = 25
