+++ play-mailer/src/main/scala/play/api/libs/mailer/Attachment.scala
