+++ play-mailer/src/main/scala/play/api/libs/mailer/Email.scala
