+++ play-mailer/src/main/scala/play/api/libs/mailer/MockMailer.scala
