+++ samples/compile-timeDI/build.sbt
@@ -5,10 +5,10 @@ name := "compile-time-DI"
 
 version := "1.0-SNAPSHOT"
 
-scalaVersion := "2.12.1"
+scalaVersion := "2.12.2"
 
 libraryDependencies ++= Seq(
-  "com.typesafe.play" %% "play-mailer" % "5.0.1-SNAPSHOT"
+  "com.typesafe.play" %% "play-mailer" % "6.0.0-SNAPSHOT"
 )
 
 lazy val root = (project in file(".")).enablePlugins(PlayScala)
